server/requirements.txt
fastapi==0.111.0
SQLAlchemy==2.0.30
GeoAlchemy2==0.15.1
alembic==1.13.1
psycopg2==2.9.9
pillow==10.3.0
